Feature/bugfixed (#58)
* remove bugs

* add upsert in seeder

* Rename Registration.blade.php to registration.blade.php

* resolve conflict

* resolve conflicts

---------

+++ database/migrations/2024_01_18_053821_remove_column_from_companies.php
@@ -13,8 +13,12 @@ return new class extends Migration
     {
         Schema::table('companies', function (Blueprint $table) {
             //
-            $table->dropColumn('username');
-            $table->dropColumn('password');
+            if (Schema::hasColumn('companies', 'username')) {
+                $table->dropColumn('username');
+            }
+            if (Schema::hasColumn('companies', 'password')) {
+                $table->dropColumn('password');
+            }
         });
     }
 
@@ -25,8 +29,12 @@ return new class extends Migration
     {
         Schema::table('companies', function (Blueprint $table) {
             //
-            $table->string('username');
-            $table->string('password');
+            if (!Schema::hasColumn('companies', 'username')) {
+                $table->string('username');
+            }
+            if (!Schema::hasColumn('companies', 'password')) {
+                $table->string('password');
+            }
         });
     }
 };
