+++ database/migrations/2024_01_25_122913_add_column_to_post_jobs_table.php
@@ -13,7 +13,9 @@ return new class extends Migration
     {
         Schema::table('post_jobs', function (Blueprint $table) {
             //
-            $table->string('skill')->after('title');
+            if (!Schema::hasColumn('post_jobs', 'skill')) {
+                $table->string('skill')->after('title');
+            }
         });
     }
 
@@ -24,7 +26,9 @@ return new class extends Migration
     {
         Schema::table('post_jobs', function (Blueprint $table) {
             //
-            $table->dropColumn('skill');
+            if (Schema::hasColumn('post_jobs', 'skill')) {
+                $table->dropColumn('skill');
+            }
         });
     }
 };
