+++ database/migrations/2024_02_26_045038_add_coumn_next_plan_id_and_subscription_id_in_user_subscription.php
@@ -13,9 +13,13 @@ return new class extends Migration
     {
         Schema::table('user_subscription', function (Blueprint $table) {
             //
-            $table->unsignedBigInteger('next_plan_id')->nullable();
-            $table->foreign('next_plan_id')->references('id')->on('plans');
-            $table->string('subscription_id')->nullable();
+            if (!Schema::hasColumn('user_subscription', 'next_plan_id')) {
+                $table->unsignedBigInteger('next_plan_id')->nullable();
+                $table->foreign('next_plan_id')->references('id')->on('plans');
+            }
+            if (!Schema::hasColumn('user_subscription', 'subscription_id')) {
+                $table->string('subscription_id')->nullable();
+            }
         });
     }
 
@@ -26,8 +30,13 @@ return new class extends Migration
     {
         Schema::table('user_subscription', function (Blueprint $table) {
             //
-            $table->dropColumn('next_plan_id');
-            $table->dropColumn('subscription_id');
+            if (Schema::hasColumn('user_subscription', 'next_plan_id')) {
+                $table->dropForeign(['next_plan_id']);
+                $table->dropColumn('next_plan_id');
+            }
+            if (Schema::hasColumn('user_subscription', 'subscription_id')) {
+                $table->dropColumn('subscription_id');
+            }
         });
     }
 };
